requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
